upgrade webrtc && support video

+++ voipengine/voipsdk/webrtc/src/webrtc/modules/audio_coding/neteq/delay_manager.h
@@ -16,7 +16,7 @@
 #include <vector>
 
 #include "webrtc/base/constructormagic.h"
-#include "webrtc/modules/audio_coding/neteq/interface/audio_decoder.h"
+#include "webrtc/modules/audio_coding/neteq/audio_decoder_impl.h"
 #include "webrtc/typedefs.h"
 
 namespace webrtc {
@@ -32,7 +32,7 @@ class DelayManager {
   // buffer can hold no more than |max_packets_in_buffer| packets (i.e., this
   // is the number of packet slots in the buffer). Supply a PeakDetector
   // object to the DelayManager.
-  DelayManager(int max_packets_in_buffer, DelayPeakDetector* peak_detector);
+  DelayManager(size_t max_packets_in_buffer, DelayPeakDetector* peak_detector);
 
   virtual ~DelayManager();
 
@@ -132,7 +132,7 @@ class DelayManager {
   void LimitTargetLevel();
 
   bool first_packet_received_;
-  const int max_packets_in_buffer_;  // Capacity of the packet buffer.
+  const size_t max_packets_in_buffer_;  // Capacity of the packet buffer.
   IATVector iat_vector_;  // Histogram of inter-arrival times.
   int iat_factor_;  // Forgetting factor for updating the IAT histogram (Q15).
   int packet_iat_count_ms_;  // Milliseconds elapsed since last packet.
